requirements.txt
flask==2.3.3
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
plotly==5.18.0
python-dateutil==2.8.2
matplotlib==3.8.2
flask-wtf==1.2.1
sqlalchemy==2.0.23
pytest==7.4.3
